+++ packages/adapter-mongoose/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone-alpha/adapter-mongoose
+
 # @voussoir/adapter-mongoose
 
 ## 2.0.1